backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
itsdangerous>=2.1.2
python-multipart>=0.0.6
starlette>=0.35.0
